+++ src/types/AppLovinMAX.ts
@@ -87,15 +87,6 @@ export type AppLovinMAXType = {
     setExtraParameter(key: string, value: string | null): void;
 
     /**
-     * @deprecated Use {@link setTermsAndPrivacyPolicyFlowEnabled()} instead.
-     *
-     * Enables the MAX Terms Flow.
-     *
-     * @param enabled true to enable the MAX Terms Flow.
-     */
-    setConsentFlowEnabled(enabled: boolean): void;
-
-    /**
      * Enables the MAX Terms and Privacy Policy Flow.
      *
      * @param enabled true to enable the MAX Terms and Privacy Policy Flow.